[website] use amplitude device id as user token for algolia
